#91 GetMeUser 관련 코드 feature_journey -> feature_main 모듈로 이동
