+++ feature_main/src/main/java/com/jyp/main/di/MyJourneyModule.kt
@@ -1,0 +1,15 @@
+package com.jyp.main.di
+
+import com.jyp.main.data.UserRepositoryImpl
+import com.jyp.main.domain.UserRepository
+import dagger.Binds
+import dagger.Module
+import dagger.hilt.InstallIn
+import dagger.hilt.android.components.ViewModelComponent
+
+@Module
+@InstallIn(ViewModelComponent::class)
+abstract class MyJourneyModule {
+    @Binds
+    abstract fun bindUserRepository(impl: UserRepositoryImpl): UserRepository
+}
