+++ feature_main/src/main/java/com/jyp/main/domain/GetMeUseCase.kt
@@ -1,4 +1,4 @@
-package com.jyp.feature_my_journey.domain
+package com.jyp.main.domain
 
 import com.jyp.core_network.jyp.ApiResult
 import com.jyp.core_network.jyp.apiResult
